fix: create secret on install and namespace
